requirements.txt
bscscan-python==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
zope.interface==5.2.0
